requirements-min.txt
scipy
matplotlib
pandas
seaborn
statsmodels
scikit-learn
nilearn
mne
pybids
imbalanced-learn
joblib
xlrd
tqdm
h5py
pymvpa2
numpy>=1.18.5
